Add per project and per scene javascript preloading hook
If a file called scene\ name.prejs is found this javascript will be evaluated before
loading the scene, the video player sample uses this to ensure that ClutterGST has
been initialized.

Another file called cluttersmith.prejs is checked for in the project root when the
project is first opened. This can be used for globally initializing a library, loading
a specific CSS style sheet or similar.